QCSQ can return values outside of i8
-129 or fewer decibells can occur

+++ common/src/status.rs
@@ -37,10 +37,10 @@ impl MiniCallHome {
         }
     }
 
-    pub fn set_signal_info(&mut self, snr_cb: i16, mut rssi_dbm: i8, rsrp_dbm: i8, rsrq_dbm: i8) {
+    pub fn set_signal_info(&mut self, snr_cb: i16, mut rssi_dbm: i8, rsrp_dbm: i16, rsrq_dbm: i8) {
         self.snr_cb = Some(snr_cb);
         if rssi_dbm == 0 {
-            rssi_dbm = rsrp_dbm - rsrq_dbm;
+            rssi_dbm = (rsrp_dbm - i16::from(rsrq_dbm)) as i8; // TODO: error if not i8
         }
         self.rssi_dbm = Some(rssi_dbm);
     }
