+++ nrf52840/src/status.rs
@@ -63,12 +63,12 @@ impl<S: Temp, T: Tx> BG77<S, T> {
         Ok((volt, bcs))
     }
 
-    async fn signal_info(&mut self) -> Result<(i8, i8, u8, i8), Error> {
+    async fn signal_info(&mut self) -> Result<(i8, i16, u8, i8), Error> {
         let response = self.simple_call_at("+QCSQ", None).await?;
         if response.count_response_values() != Ok(5) {
             return Err(Error::NetworkRegistrationError);
         }
-        Ok(response.parse4::<i8, i8, u8, i8>([1, 2, 3, 4])?)
+        Ok(response.parse4::<i8, i16, u8, i8>([1, 2, 3, 4])?)
     }
 
     async fn cellid(&mut self) -> Result<u32, Error> {
